Updated sync_prog_v2.0 to the new sync_prog_v2.1
This update removes the initial synchronization check with
the sigma repository on https://github.com/Neo23x0/sigma
and will instead pull all updates everyday.
The check with the SOC GitLab is still in place, so the
script will not attempt to pull/push new updates between
the SOC GitLab and the KIBANA Server unless it is required.

This version also keeps the "Merge Conflict Resolution
Required" email between the GitLab and the Kibana Server files.